Fixed bugs, revorked weapon:magic numbers calculation, added dummy unit
Weapon:magic now deals damage based on time which was spent to charge the weapon, rather than on charge state. Also energy consumption now depends from damage calculated for projectile.
Added a dummy unit for testing the player's weapon system

+++ Assets/Scripts/Items/EnergyPotion.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public class EnergyPotion : Item
 {
-    [SerializeField] private int energyRegenSmall = 15;
-    [SerializeField] private int energyRegenMedium = 30;
-    [SerializeField] private int energyRegenBig = 50;
+    [SerializeField] private float energyRegenSmall = 15f;
+    [SerializeField] private float energyRegenMedium = 30f;
+    [SerializeField] private float energyRegenBig = 50f;
     public override void Use()
     {
         switch (itemData.itemType)
